addExample
